refactor(core): relocate port-referenced DTOs to me-types (#816)
E.4b Phase B of the Wave 2 core decomposition. The me-storage port trait
signatures (`FactGraph`/`ConsolidationStore`/`SchemaManager`) name DTOs that
physically lived in L3/L4 modules — blocking the port from being a clean L1
leaf. Relocate those DTOs down into `me-types` (L0) and re-export them from
their old module paths so every `crate::<oldpath>::<T>` reference keeps
resolving (the `1e8600e` carve mechanic; fields widened `pub` for cross-crate
access, per the keystone precedent).

Moved (all pure data):
- `forgetting::PruneStats` -> `me_types::types::forgetting` (`prune_atomic`
  returns it). `ForgetPolicy` STAYS in `forgetting` — not port-referenced, and
  the shipped 5-method `CycleCtx` doesn't name it (plan line 575 + the DAG
  disambiguate XR-1's letter).
- `consolidation::{Snapshot, ConsolidationPlan}` + `dedup::{DedupComputed,
  Expiry}` -> `me_types::types::consolidation` (the #409 read/apply seam data;
  plan §M1).
- `archive::ArchiveManifestEntry` -> `me_types::types::archive` (archive
  feature; the monolith `archive` -> `me-types/archive` forwarding is
  pre-existing from E.3).
- `inspect::{EngineStatistics + FactStats/EdgeStats/SummaryStats/ScopeStats/
  EventStats/StorageStats, DumpFormat}` -> `me_types::types::inspect`
  (`SchemaManager::statistics`/`dump_state` reference them; plan-missed, added).

Two compiler-forced consumer updates (not behavior changes): `forgetting/policy.rs`
imports `PruneStats` from `me_types`; `storage/sqlite/schema.rs` gains a `_ =>`
arm on the now cross-crate `#[non_exhaustive]` `DumpFormat` match (unreachable
today — all four variants covered; same ripple as E.3's `CycleError::UnsupportedDelta`).

Gate: fmt + clippy --all-features -D warnings + build --all-features + test
--workspace --all-features (1967 passed, 0 failed) all green; anti-#903 test
count = 1643.
