+++ memory-engine/lib/memory-engine/src/consolidation/dedup.rs
@@ -42,58 +42,10 @@ enum DedupOutcome {
     },
 }
 
-/// One planned expiry: the `loser` to expire, folded into its `survivor`.
-///
-/// The pairing is what lets [`apply_dedup`] honor the #409 read→write gap: if the
-/// `survivor` was concurrently expired between the lock-free snapshot and the write, the
-/// merge decision is void and the `loser` must be kept (not expired), so the duplicate
-/// group is never left without an active representative. `survivor` is the **immediate**
-/// survivor of the pairwise decision; chain correctness comes from snapshotting every
-/// survivor's liveness in `apply_dedup` *before* any expiry (see there).
-pub(super) struct Expiry {
-    /// The lower-importance fact to expire.
-    pub(super) loser: i64,
-    /// The fact `loser` was folded into (the higher-importance member of the pair).
-    pub(super) survivor: i64,
-}
-
-/// Result of the pure dedup computation: which facts to expire (with their survivors) and
-/// which importance values the survivors inherit, expressed as **data** rather than DB
-/// writes.
-///
-/// Produced by [`compute_dedup`] (no `Connection`, so it runs lock-free during the
-/// engine's compute phase, #409) and applied by [`apply_dedup`] inside the final
-/// write transaction. Base-`importance` inheritance is carried separately from
-/// `importance_score`: under the current expiry rule (always expire the
-/// lower-importance fact) the base list is always empty, but it is materialized
-/// through the same guard so the applied write set is provably identical to the old
-/// inline pass.
-pub(super) struct DedupComputed {
-    /// The corpus exceeded the cap, so the pass was skipped: no writes, and the
-    /// caller must NOT advance the watermark.
-    pub(super) skipped: bool,
-    /// Planned expiries (loser + its survivor), the lower-importance member of each
-    /// duplicate pair paired with the fact it merges into.
-    pub(super) expirations: Vec<Expiry>,
-    /// Base-`importance` inheritances `(survivor_id, importance)` — empty under the
-    /// current expiry rule (kept for write-set fidelity; see the struct docs).
-    pub(super) importance_updates: Vec<(i64, f64)>,
-    /// `importance_score` inheritances `(survivor_id, score)` — the #264 running-max,
-    /// one final entry per survivor that absorbed a strictly-higher score.
-    pub(super) importance_score_updates: Vec<(i64, f64)>,
-}
-
-impl DedupComputed {
-    /// The skipped (over-cap) result: no writes, watermark held.
-    pub(super) const fn skipped() -> Self {
-        Self {
-            skipped: true,
-            expirations: Vec::new(),
-            importance_updates: Vec::new(),
-            importance_score_updates: Vec::new(),
-        }
-    }
-}
+/// `Expiry` and `DedupComputed` moved to `me-types` (Wave 2 #816 E.4b Phase B) as pure
+/// data; re-exported here so `dedup::{Expiry, DedupComputed}` keep resolving at their
+/// original in-crate path.
+pub(super) use me_types::types::consolidation::{DedupComputed, Expiry};
 
 /// Compute the local-dedup plan **without touching the store** (#409).
 ///
